move radio test driver, start initializing radio loops
